+++ BlogMvcCore/Controllers/UserController.cs
@@ -8,51 +8,19 @@ namespace BlogMvcCore.Controllers
 {
     public class UserController : Controller
     {
-        public readonly Authentication authentication;
         public readonly UserService userService;
         public readonly PostService postService;
         public readonly CommentService commentService;
-        public UserController(Authentication authentication,
-                              UserService userService,
+        public UserController(UserService userService,
                               PostService postService,
                               CommentService commentService)
         {
-            this.authentication = authentication;
             this.userService = userService;
             this.postService = postService;
             this.commentService = commentService;
         }
 
-        public IActionResult Index() => View();
 
-        public IActionResult Register() => View();
-
-        [HttpPost]
-        public IActionResult CheckRegister(string first, string second, string login,
-                                           string password, string repPassword)
-        {
-            var registrationCheck = authentication.CheckUserRegistration(first, second, login, password, repPassword);
-            return registrationCheck ? RedirectToAction("SignIn") : RedirectToAction("Register");
-        }
-
-        public IActionResult SignIn() => View();
-
-        public new IActionResult SignOut()
-        {
-            SessionHelper.SetUserAsJson(HttpContext.Session, "user", null);
-            authentication.SignOut();
-
-            return RedirectToAction("Index");
-        }
-
-        [HttpPost]
-        public IActionResult CheckIn(string login, string password)
-        {
-            var user = authentication.CheckIn(login, password);
-            SessionHelper.SetUserAsJson(HttpContext.Session, "user", user);
-
-            return user == null ? RedirectToAction("SignIn") : RedirectToAction("UserPage");
-        }
 
         public IActionResult VisitUserPage(string login)
         {
@@ -63,7 +31,6 @@ namespace BlogMvcCore.Controllers
         public IActionResult UserPage()
         {
             var userLogin = SessionHelper.GetUserFromJson<User>(HttpContext.Session, "user").Login;
-
             return View(userService.VisitUserPage(userLogin));
         }
 
@@ -74,9 +41,7 @@ namespace BlogMvcCore.Controllers
         [HttpPost]
         public IActionResult AddPost(string title, string postText, string ownerLogin)
         {
-            if (authentication.CheckStringParams(title, postText, ownerLogin))
-                postService.AddPost(title, postText, ownerLogin);
-
+            postService.AddPost(title, postText, ownerLogin);
             return RedirectToAction("UserPage");
         }
         [HttpDelete]
@@ -92,8 +57,7 @@ namespace BlogMvcCore.Controllers
         public IActionResult AddComment(string commentText, long postID, long parentID)
         {
             User user = SessionHelper.GetUserFromJson<User>(HttpContext.Session, "user");
-            if (authentication.CheckStringParams(commentText))
-                commentService.AddComment(commentText, postID, parentID, user);
+            commentService.AddComment(commentText, postID, parentID, user);
 
             return RedirectToAction("ViewPostAndComments", new { postID });
         }
@@ -107,9 +71,7 @@ namespace BlogMvcCore.Controllers
 
         public IActionResult UpdateComment(long commentID, string text, long postID)
         {
-            if (authentication.CheckStringParams(text))
-                commentService.UpdateComment(commentID, text);
-
+            commentService.UpdateComment(commentID, text);
             return RedirectToAction("ViewPostAndComments", new { postID });
         }
