+++ BlogMvcCore/Views/Authentication/Index.cshtml
